+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.23.0 (unreleased)
 
+- Changed `device` method to return `Device` instead of string
 - Fixed `item` method for multi-dimensional tensors
 
 ## 0.22.2 (2025-11-04)